requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-htmx>=1.17
django-filter>=23.0
channels>=4.0
daphne>=4.0
Pillow>=10.0
python-dotenv>=1.0
whitenoise>=6.5
django-tinymce>=3.6
stripe>=8.0
django-storages>=1.14
boto3>=1.34
django-axes>=6.1
django-ratelimit>=4.1
celery>=5.3
redis>=5.0
sentry-sdk[django]>=2.0
xhtml2pdf>=0.2.11
bleach>=6.0
python-magic-bin>=0.4
python-json-logger>=2.0
